Refactored HoardGeneratorViewModel and its factory to pass repository from fragment for use with LootGeneratorAsync. Began implementing generateHoard() on HoardGeneratorViewModel and createHoardFromOrder() on LootGeneratorAsync. Implemented creation functions on LootGeneratorAsync (still pending updated Daos and repo functions). Dummied out repo Dao implementations, pending refactor of HMRoom.kt, HMDatabaseEntries.kt, and model classes refactoring. Commit before Room refactor.

+++ app/src/main/java/com/example/android/treasurefactory/model/Hoard.kt
@@ -30,7 +30,7 @@ data class Hoard(@PrimaryKey(autoGenerate = true) val hoardID: Int = 0,
                  var spellsCount: Int = 0,
                  var isFavorite: Boolean = false,
                  var isNew: Boolean = true,
-                 var appVersion: Long = 0L,           // Version code of app hoard was generated on
+                 var appVersion: Int = 0,           // Version code of app hoard was generated on
                  @Embedded(prefix = "leftover_") val leftover: HoardLeftover = HoardLeftover())
 
 data class HoardLeftover(val gems: Int = 0,
